refactor saving into function
