+++ src/utils.py
@@ -10,6 +10,10 @@ from pathlib import Path
 from tqdm import tqdm
 import warnings
 import os 
+import csv 
+
+import pyarrow as pa 
+import pyarrow.parquet as pq
 
 
 
@@ -170,6 +174,50 @@ def check_layer_edge_dict(layer_edge_dict: Dict):
         for layer in edge_list.values():
             if len(layer) == 0:
                 raise RuntimeError("Found empty edge list")
+
+
+
+def save_to_file(data: list[list], filename: str, format="parquet") -> None:
+    """Save a list of list to parquet or csv."""
+    BATCH_SIZE_PARQUET = 100_000
+    sample_walk = data[0][0]
+    sample_walk_len = len(sample_walk)
+
+    if format == "csv":
+        with Path(filename + ".csv").open("w") as csv_file:
+            writer = csv.writer(csv_file, delimiter=",")
+            header_row = ["SOURCE"] + ["STEP_" + str(i) for i in range(sample_walk_len-1)]
+            writer.writerow(header_row)
+            for result_list in tqdm(data, desc="Writing csv"):
+                writer.writerows(result_list)
+
+    else:
+        field_col0 = [pa.field("SOURCE", pa.int64())] 
+        other_fields = [pa.field(f"STEP_{i}", pa.int64()) for i in range(sample_walk_len - 1)]
+        fields = field_col0 + other_fields
+        schema = pa.schema(fields)
+
+        def data_generator():
+            for result_list in data:
+                for result in result_list:
+                    yield {"SOURCE": result[0], **{f"STEP_{i}": step for i, step in enumerate(result[1:])}}
+
+        with pq.ParquetWriter(filename + ".parquet", schema) as writer:
+            batch = []
+            for row in tqdm(data_generator(), "Writing to parquet"):
+                batch.append(row)
+                if len(batch) >= BATCH_SIZE_PARQUET:
+                    table = pa.Table.from_pylist(batch, schema=schema)
+                    writer.write_table(table)
+                    batch.clear()
+
+            if batch:
+                table = pa.Table.from_pylist(batch, schema=schema)
+                writer.write_table(table)
+
+
+
+
